Updated to current Blender to fix various sculpt brushes.
The original mesh was created in an older version of Blender that didn't have
the Clay Strips brush. For some reason even if you opened the file in a current
version of Blender you still wouldn't have that brush as default.

To fix I just copy pasted the mesh into a fresh new file.